+++ src/main/kotlin/org/dandelion/classic/network/PacketRegistry.kt
@@ -55,6 +55,9 @@ object PacketRegistry {
         addCPE("MessageTypes")
         addCPE("InstantMOTD")
         addCPE("LongerMessages")
+        addCPE("BlockDefinitions")
+        addCPE("BlockDefinitionsExt", 2)
+        addCPE("FatMap")
     }
 
     fun registerPacket(id: Byte, factory : () -> Packet){