+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## features
 
-- added new editable message system for better server configuratoin
+- added new editable message system for better server configuration, now every server message is editable
 - improved compatibility when client does not support specific cpe extensions
 - improve disconection handling in case of excerption on connection
+- added custom blcoks support, uses mcGalaxy definition for compatibility
+- added /blocks command to manager blocks
+- 
+**Breaking changess**
+- yamlConfig changed to YamlParser
 
 ## Added CPE:
--  
+- 
+